[OF-1769] refac: External service proxy system
Migrating all external service proxy functionality out
of the user system and into its own system as a first-class
concept.
