+++ Library/src/Systems/Users/UserSystem.cpp
@@ -25,8 +25,6 @@
 #include "Common/Convert.h"
 #include "Common/UUIDGenerator.h"
 #include "Multiplayer/NetworkEventSerialisation.h"
-#include "Services/AggregationService/Api.h"
-#include "Services/AggregationService/Dto.h"
 #include "Services/UserService/Api.h"
 #include "Systems/ResultHelpers.h"
 #include "Systems/Users/Authentication.h"
@@ -35,12 +33,9 @@
 #include <regex>
 
 namespace chs_user = csp::services::generated::userservice;
-namespace chs_aggregation = csp::services::generated::aggregationservice;
 
 namespace
 {
-
-inline const char* BoolToApiString(bool Val) { return Val ? "true" : "false"; }
 
 } // namespace
 
@@ -182,7 +177,6 @@ UserSystem::UserSystem()
     , AuthenticationAPI(nullptr)
     , ProfileAPI(nullptr)
     , PingAPI(nullptr)
-    , ExternalServiceProxyApi(nullptr)
     , StripeAPI { nullptr }
     , Auth { AuthenticationAPI, CurrentLoginState }
 {
@@ -193,7 +187,6 @@ UserSystem::UserSystem(csp::web::WebClient* InWebClient, csp::multiplayer::Netwo
     , AuthenticationAPI { new chs_user::AuthenticationApi(InWebClient) }
     , ProfileAPI { new chs_user::ProfileApi(InWebClient) }
     , PingAPI { new chs_user::PingApi(InWebClient) }
-    , ExternalServiceProxyApi { new chs_aggregation::ExternalServiceProxyApi(InWebClient) }
     , StripeAPI { new chs_user::StripeApi(InWebClient) }
     , RefreshTokenChangedCallback(nullptr)
     , Auth { AuthenticationAPI, CurrentLoginState }
@@ -205,7 +198,6 @@ UserSystem::~UserSystem()
     delete (PingAPI);
     delete (ProfileAPI);
     delete (AuthenticationAPI);
-    delete (ExternalServiceProxyApi);
     delete (StripeAPI);
 
     DeregisterSystemCallback();
@@ -854,45 +846,6 @@ void UserSystem::Ping(NullResultCallback Callback)
     csp::services::ResponseHandlerPtr PingResponseHandler
         = PingAPI->CreateHandler<NullResultCallback, NullResult, void, csp::services::NullDto>(Callback, nullptr);
     static_cast<chs_user::PingApi*>(PingAPI)->pingGet({}, PingResponseHandler);
-}
-
-void UserSystem::GetAgoraUserToken(const AgoraUserTokenParams& Params, StringResultCallback Callback)
-{
-    auto TokenInfo = std::make_shared<chs_aggregation::ServiceRequest>();
-    TokenInfo->SetServiceName("Agora");
-    TokenInfo->SetOperationName("getUserToken");
-    TokenInfo->SetHelp(false);
-
-    std::map<csp::common::String, csp::common::String> Parameters;
-    Parameters["userId"] = Params.AgoraUserId;
-    Parameters["channelName"] = Params.ChannelName;
-    Parameters["referenceId"] = Params.ReferenceId;
-    Parameters["lifespan"] = std::to_string(Params.Lifespan).c_str();
-    Parameters["readOnly"] = BoolToApiString(Params.ReadOnly);
-    Parameters["shareAudio"] = BoolToApiString(Params.ShareAudio);
-    Parameters["shareVideo"] = BoolToApiString(Params.ShareVideo);
-    Parameters["shareScreen"] = BoolToApiString(Params.ShareScreen);
-
-    TokenInfo->SetParameters(Parameters);
-
-    csp::services::ResponseHandlerPtr ResponseHandler
-        = ExternalServiceProxyApi->CreateHandler<StringResultCallback, AgoraUserTokenResult, void, chs_aggregation::ServiceResponse>(
-            Callback, nullptr);
-    static_cast<chs_aggregation::ExternalServiceProxyApi*>(ExternalServiceProxyApi)->service_proxyPost({ TokenInfo }, ResponseHandler);
-}
-
-void UserSystem::PostServiceProxy(const TokenInfoParams& Params, StringResultCallback Callback)
-{
-    auto TokenInfo = std::make_shared<chs_aggregation::ServiceRequest>();
-    TokenInfo->SetServiceName(Params.ServiceName);
-    TokenInfo->SetOperationName(Params.OperationName);
-    TokenInfo->SetHelp(Params.SetHelp);
-    TokenInfo->SetParameters(Convert(Params.Parameters));
-
-    csp::services::ResponseHandlerPtr ResponseHandler
-        = ExternalServiceProxyApi->CreateHandler<StringResultCallback, PostServiceProxyResult, void, chs_aggregation::ServiceResponse>(
-            Callback, nullptr);
-    static_cast<chs_aggregation::ExternalServiceProxyApi*>(ExternalServiceProxyApi)->service_proxyPost({ TokenInfo }, ResponseHandler);
 }
 
 void UserSystem::ResendVerificationEmail(
