+++ Tests/src/PublicAPITests/ExternalServicesTests.cpp
@@ -1,0 +1,169 @@
+/*
+ * Copyright 2023 Magnopus LLC
+
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+#include "Awaitable.h"
+#include "CSP/CSPFoundation.h"
+#include "CSP/Systems/ExternalServices/ExternalServiceProxySystem.h"
+#include "CSP/Systems/SystemsManager.h"
+#include "SpaceSystemTestHelpers.h"
+#include "TestHelpers.h"
+#include "UserSystemTestHelpers.h"
+
+#include "gtest/gtest.h"
+
+namespace
+{
+
+bool RequestPredicate(const csp::systems::ResultBase& Result) { return Result.GetResultCode() != csp::systems::EResultCode::InProgress; }
+
+} // namespace
+
+CSP_PUBLIC_TEST(CSPEngine, ExternalServicesProxySystemTests, GetAgoraUserTokenTest)
+{
+    SetRandSeed();
+
+    auto& SystemsManager = csp::systems::SystemsManager::Get();
+    auto* UserSystem = SystemsManager.GetUserSystem();
+    auto* ExternlServiceProxySystem = SystemsManager.GetExternalServicesProxySystem();
+    auto* SpaceSystem = SystemsManager.GetSpaceSystem();
+
+    const char* TestSpaceName = "CSP-UNITTEST-SPACE-MAG";
+    const char* TestSpaceDescription = "CSP-UNITTEST-SPACEDESC-MAG";
+    const char* TestAssetCollectionName = "CSP-UNITTEST-ASSETCOLLECTION-MAG";
+
+    char UniqueSpaceName[256];
+    SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueString().c_str());
+
+    char UniqueAssetCollectionName[256];
+    SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueString().c_str());
+
+    auto& LogSystem = *SystemsManager.GetLogSystem();
+    bool LogConfirmed1 = false, LogConfirmed2 = false;
+    csp::common::String TestMsg1, TestMsg2;
+
+    // Log in
+    csp::common::String UserId;
+    LogInAsNewTestUser(UserSystem, UserId);
+
+    // Create space
+    csp::systems::Space Space;
+    CreateSpace(
+        SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, nullptr, Space);
+
+    // Set up Log callbacks
+    LogSystem.SetLogCallback([&](csp::common::String InMessage) { LogConfirmed1 = InMessage == TestMsg1; });
+    LogSystem.SetLogCallback([&](csp::common::String InMessage) { LogConfirmed2 = InMessage == TestMsg2; });
+
+    TestMsg1 = "AgoraUserTokenResult invalid";
+    TestMsg2 = "AgoraUserTokenResult doesn't contain expected member: token";
+
+    csp::systems::AgoraUserTokenParams Params;
+    Params.AgoraUserId = UserId;
+    Params.ChannelName = Space.Id;
+    Params.ReferenceId = Space.Id;
+    Params.Lifespan = 10000;
+    Params.ShareAudio = true;
+    Params.ShareScreen = false;
+    Params.ShareVideo = false;
+    Params.ReadOnly = false;
+
+    // Get token
+    auto [Result] = AWAIT_PRE(ExternlServiceProxySystem, GetAgoraUserToken, RequestPredicate, Params);
+
+    EXPECT_EQ(Result.GetResultCode(), csp::systems::EResultCode::Success);
+    EXPECT_FALSE(Result.GetValue().IsEmpty());
+    WaitForCallback(LogConfirmed1);
+    EXPECT_FALSE(LogConfirmed1);
+    WaitForCallback(LogConfirmed2);
+    EXPECT_FALSE(LogConfirmed2);
+
+    LogSystem.ClearAllCallbacks();
+
+    // Delete space
+    DeleteSpace(SpaceSystem, Space.Id);
+
+    // Log out
+    LogOut(UserSystem);
+}
+
+CSP_PUBLIC_TEST(CSPEngine, ExternalServicesProxySystemTests, PostServiceProxyTest)
+{
+    SetRandSeed();
+
+    auto& SystemsManager = csp::systems::SystemsManager::Get();
+    auto* UserSystem = SystemsManager.GetUserSystem();
+    auto* ExternlServiceProxySystem = SystemsManager.GetExternalServicesProxySystem();
+    auto* SpaceSystem = SystemsManager.GetSpaceSystem();
+
+    const char* TestSpaceName = "CSP-UNITTEST-SPACE-MAG";
+    const char* TestSpaceDescription = "CSP-UNITTEST-SPACEDESC-MAG";
+    const char* TestAssetCollectionName = "CSP-UNITTEST-ASSETCOLLECTION-MAG";
+
+    char UniqueSpaceName[256];
+    SPRINTF(UniqueSpaceName, "%s-%s", TestSpaceName, GetUniqueString().c_str());
+
+    char UniqueAssetCollectionName[256];
+    SPRINTF(UniqueAssetCollectionName, "%s-%s", TestAssetCollectionName, GetUniqueString().c_str());
+
+    auto& LogSystem = *SystemsManager.GetLogSystem();
+    bool LogConfirmed1 = false, LogConfirmed2 = false;
+    csp::common::String TestMsg1, TestMsg2;
+
+    // Log in
+    csp::common::String UserId;
+    LogInAsNewTestUser(UserSystem, UserId);
+
+    // Create space
+    csp::systems::Space Space;
+    CreateSpace(
+        SpaceSystem, UniqueSpaceName, TestSpaceDescription, csp::systems::SpaceAttributes::Private, nullptr, nullptr, nullptr, nullptr, Space);
+
+    // Set up Log callbacks
+    LogSystem.SetLogCallback([&](csp::common::String InMessage) { LogConfirmed1 = InMessage == TestMsg1; });
+    LogSystem.SetLogCallback([&](csp::common::String InMessage) { LogConfirmed2 = InMessage == TestMsg2; });
+    TestMsg1 = "PostServiceProxyResult invalid";
+    TestMsg2 = "PostServiceProxyResult doesn't contain expected member: token";
+
+    csp::systems::TokenInfoParams Params;
+    Params.ServiceName = "Agora";
+    Params.OperationName = "getUserToken";
+    Params.SetHelp = false;
+    Params.Parameters["userId"] = UserId;
+    Params.Parameters["channelName"] = Space.Id;
+    Params.Parameters["referenceId"] = Space.Id;
+    Params.Parameters["lifespan"] = std::to_string(10000).c_str();
+    Params.Parameters["readOnly"] = "true";
+    Params.Parameters["shareAudio"] = "false";
+    Params.Parameters["shareVideo"] = "false";
+    Params.Parameters["shareScreen"] = "false";
+
+    // Get agora token through PostServiceProxy
+    auto [Result] = AWAIT_PRE(ExternlServiceProxySystem, PostServiceProxy, RequestPredicate, Params);
+
+    EXPECT_EQ(Result.GetResultCode(), csp::systems::EResultCode::Success);
+    EXPECT_FALSE(Result.GetValue().IsEmpty());
+    WaitForCallback(LogConfirmed1);
+    EXPECT_FALSE(LogConfirmed1);
+    WaitForCallback(LogConfirmed2);
+    EXPECT_FALSE(LogConfirmed2);
+
+    LogSystem.ClearAllCallbacks();
+
+    // Delete space
+    DeleteSpace(SpaceSystem, Space.Id);
+
+    // Log out
+    LogOut(UserSystem);
+}
